Move should_elements_inline_type_behavior_be_blocked out of Document
Also change it to accept a GlobalScope, to keep it consistent with
the other methods. They refer to the same CSP List anyways.
